pyproject.toml
[tool.poetry]
name = "pigeonvision"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "pigeonvision" }, { include = "web" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
fastapi = { extras = ["standard"], version = "^0.115.12" }
python-dotenv = "^1.1.0"
idna = "^3.10"
platformdirs = "^4.3.8"
beautifulsoup4 = "^4.13.4"
jinja2 = "^3.1.6"
markdown = "^3.8"
dnspython = "^2.7.0"
grequests = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
